Terrible hacks for the RPi; needs nginx built from source
Several services are unavailable for the Raspberry Pi 3's armhf
architecture. We can ignore most of them, but we need an nginx. I
downloaded the nginx 1.10 debian source from their repository (apt-get
source is broken there) used `dpkg-source` to unpack it, and did a
`debuild -b -uc -us` and installed it. That's all the instructions you
get; good luck!