Change install instructions to use virtualenv './venv'
Explaining how to set things up with virtualenvs is easier. Using a single location is easier still. What kept me from doing this before was deciding how to clean out a location which already existed. Now check for the existence of a `$TARGET`/bin/activate.csh`. If that exists, it's pretty likely $TARGET was a virtualenv, and it won't hurt to do an install into it using `--clear`.